- Missed typo.
git-svn-id: https://svn.apache.org/repos/asf/incubator/zetacomponents/trunk@972188 13f79535-47bb-0310-9956-ffa450edef68

+++ ConsoleTools/src/input.php
@@ -402,7 +402,7 @@ class ezcConsoleInput
         {
             unset( $this->optionShort[$short] );
         }
-        if ( isset( $this->optionLong[$short] ) )
+        if ( isset( $this->optionLong[$long] ) )
         {
             unset( $this->optionLong[$long] );
         }
